Revise key expiry/genration process
New keys are generated in memory and then written to disc.  We no longer
care about which keyid we're advertising, just that we have a single
valid key. Keys also have a grace period after which they're deleted
from the in-memory secring.  A secring.new file is maintained that only
contains non-purged keys.